requirements.txt
# Core Deep Learning (Latest Versions - December 2024)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# PyTorch Geometric (for GCN)
torch-geometric==2.6.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3

# Data Processing & Visualization
folium==0.18.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84
albumentations==1.4.16

# GUI (tkinter comes with Python, no need to install separately)

# Utilities
requests==2.32.3
tqdm==4.67.1
kaggle==1.6.17
kagglehub==0.3.13

# Optional but highly recommended for development
jupyter==1.1.1
ipykernel==6.29.5

# Additional packages for PyTorch 2.5.1 compatibility
accelerate==1.1.1
transformers==4.41.2
datasets==3.1.0